* Got site working without-known-error again, in dev-mode, dev-deploy, and prod-deploy. : ) [fixed css issue, etc.] * Switched to having synthetic-default-imports disabled. (still not sure on this one; an advantage of enabled is that it's less likely to error from wrongly using module.default, disadvantage is that it's a misdirection from the real/raw state of the data (and had that issue with ts targeting es2017 -- not sure if that still applies))
Note: If site doesn't load at first on deploy-dev (and maybe the others), just try closing the tab completely and opening it in a new tab without dev-tools open. There was a strange bug where having dev-tools open and/or staying on the same tab kept causing an error to occur before the page code finished first-running.

Former-commit-id: 5312498b1606ac7945616fc5fce0efb4d4582647