Note: Relocated main repo to https://github.com/debate-map/client, and site to debatemap.app.
Because of the relocation, this repo is now the "old" version, so the homepage is updated to mention the relocation.

This version's hosting domain is still debatemap.live; however, it's hosting project is now "debate-map-prod-old". HOWEVER, its actual db data is still stored in the debate-map-prod project (it just uses the realtime-database instead of the Firestore database).

By the way, note that the dist/app.js file was manually modified to include the relocation note. This is because, the compiling process is currently broken (due to my not uploading my libs, eg. vwebapp-framework, to npm when I was developing in this repo). It's fine though since this old repo/site (debatemap.live) is only up so I can move its content over to the new site (debatemap.app), with an actual UI available.


Former-commit-id: 5e0fd434a22338a23de705f72ebe007ed7ddc27b