* Reduced the image-deploy time from ~3m to ~1s by switching from k3d/kind to Docker Desktop for the local kubernetes implementation. (you can get very fast loading in k3d/kind as well, by setting up a local registry, but I haven't learned that yet, and Docker Desktop Kubernetes seems sufficient for now)
There is still ~3m of delay when the docker build detects it needs to rerun `yarn install`. But that's not as critical, since I can "workaround it" using Skaffold sync-inclusions for all the "hot paths", eg. the debate-map source [actually, this is not so critical anyway since it is copied at a step later than yarn-install in the dockerfile], as well as that of web-vcore and my libs [presumably skaffold-sync can work for libs, anyway])