* Added tilt-down scripts to package-scripts.js. * Added back a section of the guide I had actually failed to copy over from the deploy readme. (specifically, the cloud-project-init module) * Tried switching the ovh k8s cluster to use an 8gb-memory node rather than 4gb-memory.
I've hit a number of issues along the way, but have resolved all of them except one (seemingly): the websocket connections are being immediately dropped! I can see that the data in the database is correct, and the non-websocket requests to the web-server and app-server seem to all work, but for some reason the websocket connections are not (the drops are seen in the dev-tools of Chrome, and it eventually leads to a frontend error -- no logs/issues are apparent on the backend, and the backend health checks are completing, so the issue must be between the app-server and the browser/frontend [perhaps something relating to the ingress/traefik controller]).

Not really sure where to begin debugging the issue. (I suppose by adding lots logging/debugging lines throughout the websocket-related code of postgraphile and/or the live-query plugin)