* Got basic implementation of async_await::stream! + channels approach written; finally compiling without errors again.
However, it's having a race condition at runtime because the storage_wrapper.lock() guard is being kept around while the loop waits for the await_next_entries future to resolve. So I'll work on that next. (perhaps using Arc on the channel receiver, and cloning it, then sending only that into the loop)