* Implemented the proxy between app-server-rs and app-server-js; with this, the site has basic usability restored!
However, the final-state frontend performance -- while improved, and of course much more stable than before -- is not as speedy as I expected. On doing a profile, it appears the frontend just has lots of inefficiencies that are causing the data-request loop to progress too slowly. (that is, the slowdown is primarily in the frontend's slow progression through the steps required for it to know what to request next)

So I will likely focus on that next, for a day or two, as there seems to be a lot of low hanging fruit. Prime example: the "useDefineForClassFields" setting for TypeScript. I'm going to disable it, because over the profiling duration that captured a single node rename/revision-creation, the Object.defineProperty calls took a total of ~750ms. Partly this might be due to a bug in my web-vcore code (of not successfully detaching, which is easy to fix), but we'll see.