* Made some improvements to FlashKit. * Created sub-nav-bar entry for the "focus" map-viewing mode. (not yet implemented) * Started work on making-so if there is only one node expanded in a sibling-set, its descendants have free use of the space to the right. (ie. the node doesn't need to have any gap between it and the other siblings)
Quick backup before reworking layout logic, to have that right-area using position:absolute; this way, we have more control over the layout. For example, the size increase of a deep descendant does not instantly (and uncontrolled-ly) cause the entire ancestor tree to resize with it; instead, we must manually propagate the size change, thus letting us keep the reactions synchronized.