* Made some more progress on the new Command system in app-server-rs.
However, I am postponing the above, because it would take too long to reimplement all of the subcommands that are necessary for the new transfer-nodes command. Instead, for now I'll just implement it in app-server-js; then I'll gradually migrate each command over, refining and testing the Rust implementations as I go. (safer approach)