* MS the web-server, app-server, etc. are accessible from the plain-http ovh-host URLs again. * MS the dashboard is accessible at "<ovh-host>:9000". (finally... that's how I diagnosed the next point:) * Okay... well I think I figured out one of the main reasons everything's been so broken: Apparently Traefik has a wonderful behavior where if the https certificates for the second gateway are unavailable (because, you know, we're launching the server in order to generate them), then the error in that gateway's launch prevents the regular-http gateway from attaching/activating its own routes. Which... means that the route which enables the certificates themselves to be provisioned, goes down. Which means it keeps itself from getting started. :/
There must be some approach that avoids that catch-22; but the fact that something that annoying happened to begin with (without me seeing anything about that pitfall in any of my reading) brings into question some of the structural/robustness designs that were made. Maybe it's because the API is still in alpha; but if I can't find an easy solution to this, I think I might be done with Traefik.