* Made the first large step toward implementing the new node-layout approach, which no longer "merges" claims and arguments into the same node-box. (it led to too much confusion for editors)
This new approach is also different than the separate-node-box approaches that came before, since it provides a place for both the argument and claim nodes to place their toolbar-buttons "on the same row" -- thus enabling visual separation of all nodes, while not increasing the row-count/space-taken beyond the combined-node-box approach with node-toolbars enabled. (we have the argument show one "Relevance" toolbar-button to the right of itself, and we have the claim node shows its two-or-so toolbar-buttons just above itself, floated to the right; so long as the toolbar-button-count is reasonable, they thus can fit on the same row)

This new approach also makes multi-premise arguments a lot more ergonomic: it looks the same as regular arguments except it has an extra premise below the first (no extra/unexpected wrapper), and no longer requires the person to have to think of what summary text to write for that argument wrapper (which would also have been a substantial curation challenge).

At the moment, there are quite a few "hacks" in place to get the visuals how they're wanted (eg. manipulation of the IsSinglePremiseArgument and IsMultiPremiseArgument functions, and lots of hard-coded values/modifications); these will be transitioned out as the implementation process progresses.