* Tried to jerry-rig a solution to the error in traefik: `reflector.go:424] k8s.io/client-go@v0.26.3/tools/cache/reflector.go:169: failed to list *v1alpha2.XXX: the server could not find the requested resource (get XXX.gateway.networking.k8s.io)``)
Note1: The attempt seemed to work at first (with equivalent change to just 0.7.0 version of gatewayclasses.yaml seeming to avoid that error message), but when applied to all of the files (and in version 0.8.0 of crds), the warning/error message appears to still happen. (though note that the file changes *did* succeed in forcing all of the gateway CRDs to v1alpha2, which presumably should fix it since that's what the error message said it was looking for; perhaps the issue is a permissions one, eg. where traefik's reflector.go file just has insufficient permissions to find those resources)

Note2: Nevermind; after deleting and redeploying the "gateway-api-crd" and "gateway-api-other-objects" groups, and moving the gatewayclasses object to a new crds "early" group, the error message about v1alpha2 crds not being findable went away.

Note3: However, Traefik is still failing to find/initialize the gateways/routes.

I think what happened is that an earlier traefik/gateway-apis upgrade (perhaps of gateway-apis, from v0.4.3 to 0.7.0) actually had the data-files in an invalid state, but I just didn't notice, because "tilt up" did not fully flush out the old contents (perhaps the crds), leading to the live-clusters still working but the data-files not being able to support that if starting from scratch.

Assuming this is the case (which is supported by, eg. the life-times of the gateway stuff in my still-working local cluster), then perhaps a way to mitigate this occurence is to do a full "teardown" of the cluster (perhaps using tilt-down, or just ending tilt-up with a regular ctrl-c then wait rather than force-close), then do a fresh "tilt up", whenever I make major changes to the k8s files (those relating to crds anyway).