* MS Symbola font is served by the web-server, and is listed just after Quicksand; this allows many symbols to be displayed on all devices, and in a consistent way. * Organized .gitignore file a bit more nicely.
Note: The addition of Symbola font does mean an extra 889kb of data transferred, which is not great. In the long-term, will look for ways to reduce this, eg. by making a variant of Symbola that only includes its symbols/non-letters. Including Symbola seems to make sense for now though (it allows using the many variants of arrows directly in node-text, the `(a)` family of symbols commonly used in scientific papers, etc.); in any case, it's far from the only source of bloat atm. (eg. material-design icons is ~1mb, js bundle is >10mb, etc.)