* Removed the "parents count" filter in the batch "set access policy" panel for now.
Reason: I just realized that the "parents count" filter does not work currently, because the SQL "descendants" function is only returning node-links that are part of that subtree, not all node-links that include the subtree's nodes as children in general. So removing that part for now, since atm it gives a false sense of safety. (long-term I may restore it, after adding a SQL function that can provide that data efficiently)