* Fixed that rust builds were failing (at runtime), by disabling usage of the mold linker.
Note: The runtime errors *seem* to have started happening due to a new version of mold being released yesterday [v2.31.0], and getting included in the sid/unstable branch of debian's package repository (which is where we pull compiled mold from).

Somehow that ends up causing GLIBC 2.38 to be required on the OS image used for running the produced rust binary, which is too new of a version, even when switching to the latest "trixie" [ie. v13] of debian-slim.

Confusingly, the issue still happens even when installing mold from github using an old tag, so it may not be mold directly that is causing it; but it seems to at least be part of the trigger, because disabling usage of the mold linker in config.toml [even when mold was still installed in the dockerfile] fixed the runtime issue.