* Updated react-vmessagebox. Relevant changes: ** MS dialog titles don't change the sizing of the dialog. (changed them to use absolute positioning, and cut off overflow) * Subtree-deletion function in the "subtree operations" dialog is now functional! Closes #319.
Note: This implementation is fairly "naive". For example, it doesn't handle the case where a node is linked to a map only through the tree in question, but the tree includes a link to it twice. In that case, the code needs to unlink it the first time it sees it, and delete it the second time it sees it; but instead, the current implementation will simply error if this case is hit. (one of the "unlink node" commands will fail due to it being the last place the child is linked from)

I don't want to take the time to add handling for those more complex cases right now. For now, the mass-deleter is intended just for "clean" map states, where all the nodes are either "single parent", or multi-parent but with at least one link outside of the current tree being deleted.