* Added PGDumpBackupHelper.js.
This script is an alternate way of automating database backups-to-local-pc, by starting a kubernetes port-forward (if it's not active already), and having pgdump connect and save the db contents directly.

This is a bit less convenient (eg. requires sharing/storing of db's password itself, and setup of kubectl, and download of pgdump tool), but is also more "robust", since it doesn't need to pass through NGINX, the app-server, etc.